requirements.txt
streamlit
pandas
numpy
seaborn
matplotlib.pyplot
functions_app